+++ __init__.py
@@ -2,11 +2,12 @@ bl_info = {
     "name": "Ref to image plane",
     "description": "convert a reference image to image plane (textured quad mesh)",
     "author": "Samuel Bernou",
-    "version": (0, 1, 0),
+    "version": (0, 2, 0),
     "blender": (2, 93, 0),
     "location": "View3D",
     "warning": "",
-    "doc_url": "", #https://github.com/Pullusb/PROJ_name
+    "doc_url": "https://github.com/Pullusb/ref_to_image_plane",
+    "tracker_url": "https://github.com/Pullusb/gp_draw_transform/issues/new",
     "category": "Object" }
 
 from . import OP_ref_to_image_plane