+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+0.2.0
+
+- feat: option to get name from orignal image or from empty object
+- fix: bug with linking
+- doc: fic readme link and description
+- code: cleaner:
+    - copied function locally instead of importing from IAP
+    - add bug report url
+    - add git-ignore
+
 0.1.0
 
 - working version